+++ src/Fantomas.Core.Tests/ExternTests.fs
@@ -41,7 +41,12 @@ module Foo =
 module Foo =
     module Bar =
         [<DllImport("Kernel32.dll", SetLastError = true)>]
-        extern bool GetFileInformationByHandleEx(IntPtr hFile, FILE_INFO_BY_HANDLE_CLASS infoClass, [<Out>] FILE_NAME_INFO& info, uint32 dwBufferSize)
+        extern bool GetFileInformationByHandleEx(
+            IntPtr hFile,
+            FILE_INFO_BY_HANDLE_CLASS infoClass,
+            [<Out>] FILE_NAME_INFO& info,
+            uint32 dwBufferSize
+        )
 """
 
 [<Test>]
@@ -133,9 +138,13 @@ open System.Runtime.InteropServices
 open Accessibility
 
 [<DllImport("oleacc.dll")>]
-extern int AccessibleChildren(IAccessible paccContainer, int iChildStart, int cChildren, [<Out;
-                                                                                           MarshalAs(UnmanagedType.LPArray,
-                                                                                                     SizeParamIndex = 4s)>] System.Object [] rgvarChildren, int* pcObtained)
+extern int AccessibleChildren(
+    IAccessible paccContainer,
+    int iChildStart,
+    int cChildren,
+    [<Out; MarshalAs(UnmanagedType.LPArray, SizeParamIndex = 4s)>] System.Object[] rgvarChildren,
+    int* pcObtained
+)
 """
 
 [<Test>]
@@ -172,4 +181,83 @@ extern UIntPtr private GetProcessHeap()
         """
 [<DllImport("kernel32.dll")>]
 extern UIntPtr private GetProcessHeap()
+"""
+
+[<Test>]
+let ``extern C-like parameters, 1216`` () =
+    formatSourceString
+        false
+        """
+extern int M1(int)
+
+extern int M2(int i)
+
+extern int M3([<A>] int i)
+
+extern int M4(int[] i)
+
+extern int M5(int* i)
+
+extern int M5(int& i)
+
+extern int M5(void* i)
+
+extern int M6(void* i1, void* i2)
+"""
+        config
+    |> prepend newline
+    |> should
+        equal
+        """
+extern int M1(int)
+
+extern int M2(int i)
+
+extern int M3([<A>] int i)
+
+extern int M4(int[] i)
+
+extern int M5(int* i)
+
+extern int M5(int& i)
+
+extern int M5(void* i)
+
+extern int M6(void* i1, void* i2)
+"""
+
+[<Test>]
+let ``extern with void return type, 1215`` () =
+    formatSourceString
+        false
+        """
+extern void GetProcessHeap()
+extern [<A>] void GetProcessHeap2()
+"""
+        config
+    |> prepend newline
+    |> should
+        equal
+        """
+extern void GetProcessHeap()
+extern [<A>] void GetProcessHeap2()
+"""
+
+[<Test>]
+let ``extern declaration inside type declaration, 1214`` () =
+    formatSourceString
+        false
+        """
+type T() =
+  [<DllImport("kernel32.dll")>]
+  extern UIntPtr private GetProcessHeap()
+"""
+        config
+    |> prepend newline
+    |> should
+        equal
+        """
+type T() =
+    [<DllImport("kernel32.dll")>]
+    extern UIntPtr private GetProcessHeap()
 """